imported_app/AI Singapore/requirements.txt
# AI Log Parser — Semiconductor Log Pipeline
# Core AI / LLM

# OpenAI Python client
openai>=1.0.0

# Web framework (streaming endpoint)
flask>=3.0.0

# ML / clustering / anomaly detection
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0

# Data processing
pandas>=2.2.0

# Parsing formats
PyYAML>=6.0.1
lxml>=5.1.0

# Model persistence
joblib>=1.3.0

# Timestamp parsing
python-dateutil>=2.9.0

# Environment variable loading from .env file
python-dotenv>=1.0.0
